requirements.txt
tensorflow-cpu>=2.12
keras-tuner>=1.4.6
streamlit>=1.28
pillow
numpy
scikit-learn
matplotlib
